+++ src/DuckGame/Equipment/Equipment.cs
@@ -142,7 +142,7 @@ namespace DuckGame
             if (this.destroyed)
             {
                 this.alpha -= 0.1f;
-                if ((double)this.alpha < 0.0)
+                if (this.alpha < 0.0)
                     Level.Remove(this);
             }
             if (this.localEquipIndex < this.equipIndex)
