+++ src/DuckGame/FeatherVolume.cs
@@ -29,12 +29,12 @@ namespace DuckGame
             if (bullet.owner != null && (bullet.owner == this._duckOwner || owner != null && owner.owner == this._duckOwner))
                 return false;
             Feather feather = Feather.New(0f, 0f, this._duckOwner.persona);
-            feather.hSpeed = (float)(-(double)bullet.travelDirNormalized.x * (1.0 + (double)Rando.Float(1f)));
+            feather.hSpeed = (float)(-bullet.travelDirNormalized.x * (1.0 + Rando.Float(1f)));
             feather.vSpeed = -Rando.Float(2f);
             feather.position = hitPos;
             Level.Add(feather);
             Vec2 point = hitPos + bullet.travelDirNormalized * 3f;
-            if (bullet.isLocal && this._duckOwner.sliding && this._duckOwner.ragdoll == null && point.x > (double)this.left + 2.0 && point.x < (double)this.right - 2.0 && point.y > (double)this.top + 2.0 && point.y < (double)this.bottom - 2.0)
+            if (bullet.isLocal && this._duckOwner.sliding && this._duckOwner.ragdoll == null && point.x > this.left + 2.0 && point.x < this.right - 2.0 && point.y > this.top + 2.0 && point.y < this.bottom - 2.0)
             {
                 foreach (Equipment equipment in Level.CheckPointAll<Equipment>(point))
                 {
@@ -52,7 +52,7 @@ namespace DuckGame
             if (bullet.owner != null && (bullet.owner == this._duckOwner || owner != null && owner.owner == this._duckOwner))
                 return;
             Feather feather = Feather.New(0f, 0f, this._duckOwner.persona);
-            feather.hSpeed = (float)(-(double)bullet.travelDirNormalized.x * (1.0 + (double)Rando.Float(1f)));
+            feather.hSpeed = (float)(-bullet.travelDirNormalized.x * (1.0 + Rando.Float(1f)));
             feather.vSpeed = -Rando.Float(2f);
             feather.position = exitPos;
             Level.Add(feather);
